requirements.txt
flask
soco
spotipy
adafruit-circuitpython-pn532
RPi.GPIO
python-dotenv
requests
